Update theme and implement modular commands system
Let the actual real work commence, folks! This is the first time commands are working. All that's left to do with it is implement returning to the chat properly. We also need to work on logging promises. Errors are lurking around every corner!

I'd like to thank myself, myself, and me for making this possible. What a bit of Googling can't accomplish...