Update theme and improve navigation
We're now keeping the chat on the screen at all times to ensure that you are always able to interact with the chat.

Background stuff; we improved the handling of refreshing by closing any existing Irc clients that are open and throwing it away entirely. This should make things less memory-hungry. Additionally, this allows us to get dynamic content away from the background thread.